refactor(live-preview): separate scopes for libraries
