+++ packages/live-previews/src/scope/mantine.tsx
@@ -1,0 +1,62 @@
+import React from "react";
+import type { RefineProps } from "@pankod/refine-core";
+import { RefineCommonScope } from "./common";
+import * as RefineMantine from "@pankod/refine-mantine";
+
+const SIMPLE_REST_API_URL = "https://api.fake-rest.refine.dev";
+
+const RefineMantineDemo: React.FC<
+    Partial<RefineProps> & {
+        initialRoutes?: string[];
+    }
+> = ({ initialRoutes, ...rest }) => {
+    if (initialRoutes) {
+        RefineCommonScope.setInitialRoutes(initialRoutes);
+    }
+
+    return (
+        <RefineMantine.MantineProvider
+            theme={RefineMantine.LightTheme}
+            withNormalizeCSS
+            withGlobalStyles
+        >
+            <RefineMantine.Global
+                styles={{ body: { WebkitFontSmoothing: "auto" } }}
+            />
+            <RefineCommonScope.RefineCore.Refine
+                routerProvider={RefineCommonScope.RefineReactRouterV6.default}
+                dataProvider={RefineCommonScope.RefineSimpleRest.default(
+                    SIMPLE_REST_API_URL,
+                )}
+                notificationProvider={RefineMantine.notificationProvider}
+                Layout={RefineMantine.Layout}
+                Sider={() => null}
+                catchAll={<RefineMantine.ErrorComponent />}
+                options={{
+                    disableTelemetry: true,
+                    reactQuery: {
+                        devtoolConfig: false,
+                    },
+                }}
+                {...rest}
+            />
+        </RefineMantine.MantineProvider>
+    );
+};
+
+const MantineScope = {
+    // ...RefineCommonScope,
+    // RefineAntdDemo,
+    // RefineAntd,
+    // RefineMuiDemo,
+    // RefineMui,
+    RefineMantine,
+    RefineMantineDemo,
+    // RefineChakra,
+    // RefineChakraDemo,
+    // // Other Packages
+    // RefineReactHookForm,
+    // RefineReactTable,
+};
+
+export default MantineScope;
