+++ packages/live-previews/src/scope/map.tsx
@@ -1,0 +1,11 @@
+export const packageMap: Record<string, string> = {
+    "@pankod/refine-core": "RefineCore",
+    "@pankod/refine-react-router-v6": "RefineReactRouterV6",
+    "@pankod/refine-antd": "RefineAntd",
+    "@pankod/refine-mui": "RefineMui",
+    "@pankod/refine-mantine": "RefineMantine",
+    "@pankod/refine-chakra-ui": "RefineChakra",
+    "@pankod/refine-simple-rest": "RefineSimpleRest",
+    "@pankod/refine-react-hook-form": "RefineReactHookForm",
+    "@pankod/refine-react-table": "RefineReactTable",
+};
